// src/module/users/dto/create-user.dto.ts
export interface ProvisionedAccountsDTO {
  palmpay: {
    accountId?: string;
    accountNumber?: string;
    accountName?: string;
    bankName: string;
    currency: string;
  };
  pouchii: {
    walletId: number;
    accountNumber: string;
    accountName: string;
    bankName: string;
    currency: string;
  };
}
// ...
export interface PouchiiWalletResponseDTO {
  responseCode: number
  responseMessage: string
  requiredVerification: boolean
  responseTime: string
  data: PouchiiWalletResponseData
}

export interface PouchiiWalletResponseData {
  walletType: string
  walletId: number
  currency: Currency
  walletNumber: string
  walletName: string
  schemeId: string
  walletKycLevel: string
  kycWalletInfo: KycWalletInfo
  accountNumber: string
  accountName: string
  bookedBalance: number
  availableBalance: number
  creditLienAmount: number
  debitLienAmount: number
  bank: Bank
  isPrimaryWallet: boolean
  walletStatus: boolean
}

export interface Currency {
  name: string
  code: string
  symbol: string
}

export interface KycWalletInfo {
  hasPendingKycUpgradeApproval: boolean
  pendingKycUpgradeRequest: PendingKycUpgradeRequest
  requiredKycUpgradeAction: boolean
  kycUpgradeActionInfo: KycUpgradeActionInfo
}

export interface PendingKycUpgradeRequest { }

export interface KycUpgradeActionInfo {
  currentKycLevel: string
}

export interface Bank {
  id: number
  code: string
  name: string
}

export interface PalmpayBaseRequest {
  requestTime?: number;
  version?: string;
  nonceStr?: string;
}

export interface PalmpayBaseResponse {
  respCode: string;
  respMsg: string;
  status: boolean;
}
// ...
export interface CreateVirtualAccountResponseData {
  identityType: string
  licenseNumber: string
  virtualAccountName: string
  virtualAccountNo: string
  email: string
  customerName: string
  status: string
  accountReference: string
  appId: string
}

export interface CreateVirtualAccountResponse extends PalmpayBaseResponse {
  data: CreateVirtualAccountResponseData | null
}